+++ api/quizs.js
@@ -1,38 +1,81 @@
-let quizs = "This quiz is from ./server/api/quizs.js";
+// let quizs = "This quiz is from ./server/api/quizs.js";
 
 const router = require("express").Router();
+const mysql = require('mysql');
+const mysqlConnection = mysql.createPool({
+  host: 'us-cdbr-iron-east-02.cleardb.net',
+  user: "b51a3d0daa4770",
+  password: '43d9010d',
+  multipleStatements: true,
+  database: 'heroku_1b3f9d32c7d74e0'
+})
 
-//This is a get request, the route is '/api/quizs'
-//For querying all items.
-router.get("/", (req, res) => res.send(quizs));
+mysqlConnection.getConnection(err=>{
+  if(!err) console.log('DB connection succeeded!');
+  else console.log('DB Connection Error: '+ JSON.stringify(err, undefined, 2))
+  
+})
 
-//This is a get request, the route is '/api/quizs/:id'
-//For querying an item by the argument from url.
-router.get("/:id", (req, res) => {
-  //Type of 'req.params.id' is string.
-  let id = Number(req.params.id);
-  for (let i = 0; i < quizs.length; i++) {
-    if (quizs[i].id === id) res.send(quizs[i]);
-  }
-});
+router.get('/', (req, res)=>{
+  mysqlConnection.query('SELECT * FROM quiz',
+  (err, rows, field)=>{
+    if(!err){
+      res.send(rows);
+    }
+    else(console.log(err))
+  })
+})
 
-//This is a post request, the route is '/api/quizs'
-//For adding a new item.
-//should get a prameter call name which is from request.
-router.post("/", (req, res) => {
-  let newId = quizs.length + 1;
-  let newExample = {
-    id: newId,
-    name: req.body.name
-  };
-  quizs.push(newExample);
-  res.send(quizs);
-});
+ 
+//Add an new item without input QuizID
+router.post('/', (req, res) => {
+  let emp = req.body;
+  console.log(emp)
+  emp.id = 0;
+  
+  let sql = 'SET @id = ?;SET @question = ?;SET @option1 = ?;SET @option2 = ?; SET @option3 = ?; SET @option4 = ?; \
+  CALL QuizAddOrEdit(@id, @question, @option1, @option2, @option3, @option4)';
+  mysqlConnection.query(
+    sql, [emp.id, emp.question, emp.option1, emp.option2, emp.option3, emp.option4],
+    (err, rows, field) => {
+      if (!err) res.send(emp);
+      else console.log(err)
+    }
+  )
+})
+//Update an item by quizID.
+router.put('/', (req, res) => {
+  let emp = req.body;
+  
+  let sql = 'SET @id = ?;SET @question = ?;SET @option1 = ?;SET @option2 = ?; SET @option3 = ?; SET @option4 = ?; \
+  CALL QuizAddOrEdit(@id, @question, @option1, @option2, @option3, @option4)';
+  mysqlConnection.query(
+    sql, [emp.id, emp.question, emp.option1, emp.option2, emp.option3, emp.option4],
+    (err, rows, field) => {
+      if (!err) res.send('Updated successfully');
+      else console.log(err)
+    }
+  )
+})
+
+//similar to export default
+module.exports = router;
+
+
+
+
+
+
+
+
+
+/** -----------------------------THESE CODES ARE FOR REFERENCE ONLY-----------------------------------------  */
+
+
 
 //This is a put request, the route is '/api/students'
-//For updating a new item.
-//should get a prameter call id which is from url
-router.put("/:id", (req, res) => {
+//For updating a new item. should get a prameter call id which is from url
+router.put("/:idDFSADFSF", (req, res) => {
   //Type of 'req.params.id' is string
   let id = Number(req.params.id);
   let studentIndex = 0;
@@ -52,12 +95,10 @@ router.put("/:id", (req, res) => {
 //This is a post request, the route is '/api/students'
 //For deleting a new item.
 //should get a prameter call id which is from request
-router.delete("/:id", (req, res) => {
+router.delete("/:idADFSDFDSF", (req, res) => {
   //Type of 'req.params.id' is string
   let id = Number(req.params.id);
   students = students.filter(el => el.id !== id);
   res.send(students);
 });
 
-//similar to export defaul
-module.exports = router;
